Add Sponsors link to nav menus
This PR should enable a link to Sponsors to appear in the header and footer of the c4l site. 

Ideally, this link should send users to our Prospectus page first as we're not ready for listing Sponsors quite yet, but we'll find out when we merge! 😅 Not the end of the world as long as the site still builds.