Changed many files for posting schedule 📅
Many details needed to be changed to accommodate the workshops being at the end, including day numbering, as we agreed (https://code4lib.slack.com/archives/C0D9KS1EU/p1699039447250239) that we would move to 1-based indexing. To reinforce this, I chose to add a dummy date to _data/conf.yml so any indexes in templates/variables could sync with the day of the conference. Open to other ideas.